requirements.txt
APScheduler==3.9.1
beautifulsoup4==4.11.1
confluent-kafka==1.7.0
dms2dec==0.1
Flask==2.1.1
Flask-Cors==3.0.10
geopandas==0.9.0
gunicorn==20.1.0
lxml==4.8.0
marshmallow==3.15.0
opencv-python==4.6.0.66
pandas==1.3.1
patool==1.12
pdf2image==1.16.0
pdfplumber==0.7.1
pymongo==4.1.0
python-dotenv==0.20.0
pytesseract==0.3.9
requests==2.26.0
selenium==3.141.0
Unidecode==1.3.4
Werkzeug==2.1.1